% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{balance}
\alias{balance}
\title{Balance groups by up- or downsampling.}
\usage{
balance(data, size, cat_col, id_col = NULL, id_method = "n_ids",
  mark_new_rows = FALSE, new_rows_col_name = ".new_row")
}
\arguments{
\item{data}{Dataframe.}

\item{size}{Size to fix group sizes to.
 Can be a specific number, given as a whole number, or one of the following strings:
 \code{"min"}, \code{"max"}, \code{"mean"}, \code{"median"}.

 \subsection{number}{
 Fix each group to have the size of the specified number of row.
 Uses downsampling for groups with too many rows and upsampling for groups with too few rows.
 }
 \subsection{min}{
 Fix each group to have the size of smallest group in the dataset.
 Uses downsampling on all groups that have too many rows.
 }
 \subsection{max}{
 Fix each group to have the size of largest group in the dataset.
 Uses upsampling on all groups that have too few rows.
 }
 \subsection{mean}{
 Fix each group to have the mean group size in the dataset. The mean is rounded.
 Uses downsampling for groups with too many rows and upsampling for groups with too few rows.
 }
 \subsection{median}{
 Fix each group to have the median group size in the dataset. The median is rounded.
 Uses downsampling for groups with too many rows and upsampling for groups with too few rows.
 }}

\item{cat_col}{Name of categorical variable to balance by. (Character)}

\item{id_col}{Name of factor with IDs. (Character)
 This will be used to respect IDs as entities, which can only be added or removed in totality.


 E.g. If we have measured a participant multiple times and
 want make sure that we have all these measurements. Then we would either
 remove/add all measurements for the participant or leave in
 all measurements for the participant.}

\item{id_method}{Method for balancing the IDs. (Character)

 \code{n_ids} or \code{n_rows_c}.
 \subsection{n_ids}{
 Balances on ID level only. It makes sure there are the same number of IDs for each category.
 This might lead to different number of rows in categories.
 }
 \subsection{n_rows_c}{
 Attempts to level the number of rows per category, while only removing/adding entire IDs.
 This is done by:
  \subsection{Repetition}{
  If a category needs to add all its rows one or more times, the data is repeated.
  }
  \subsection{Closest}{
  Iteratively, the ID with the number of rows closest to the
  lacking/excessive number of rows is added/removed.
  This happens until adding/removing the closest ID would lead to a size further from
  the target size than the current size.
  If multiple IDs are closest, one is picked by random sampling.
  }
 }}

\item{mark_new_rows}{Add column with 1s for added rows, and 0s for original rows. (Bool)}

\item{new_rows_col_name}{Name of column marking new rows. Defaults to ".new_row".}
}
\value{
Dataframe with added and/or deleted rows. Ordered by cat_col and (potentially) id_col.
}
\description{
Uses up- or downsampling to fix the group size to the
 min, max, mean, or median group size or to a specific number of rows.
}
\details{
\subsection{Without id_col}{Upsampling is done with replacement for added rows, while the original data remains intact.
Downsampling is done without replacement, meaning that rows are not duplicated but only removed.}
\subsection{With id_col}{See id_method description.}
}
\examples{
# Attach packages
library(groupdata2)

# Create dataframe
df <- data.frame(
 "participant" = factor(c(1, 1, 2, 3, 3, 3, 3)),
 "trial" = c(1,2,1,1,2,3,4),
 "score" = sample(c(1:100), 7))

# Using balance() with number
balance(df, 3, "participant")

# Using balance() with min
balance(df, "min", "participant")

# Using balance() with max
balance(df, "max", "participant")

}
\seealso{
Other sampling functions: \code{\link{downsample}},
  \code{\link{upsample}}
}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}
