% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearrange.R
\name{center_max}
\alias{center_max}
\title{Centers the highest value with values decreasing around it}
\usage{
center_max(data, col = NULL, shuffle_members = FALSE)
}
\arguments{
\item{data}{Data frame or vector. When a vector, it is converted to a data frame.}

\item{col}{Column to create sorting factor by. When \code{NULL} and \code{data} is a data frame,
it uses the row numbers.}

\item{shuffle_members}{Whether to shuffle the pair members. (Logical)

For the \code{"center_*"} methods, this randomizes which values are
to the right and left of the center.}
}
\value{
The sorted data frame.
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}

The highest value is positioned in the middle with the other
values decreasing around it.

Wrapper for \code{\link[groupdata2:rearrange]{rearrange()}} with \code{method = "center_max"}.

\strong{Example}:

The column values:

\code{c(1, 2, 3, 4,} \strong{\code{5}}\code{)}

are \strong{ordered as}:

\code{c(1, 3,} \strong{\code{5}}\code{, 4, 2)}
}
\examples{
\donttest{
# Attach packages
library(groupdata2)

# Set seed
set.seed(1)

# Create a data frame
df <- data.frame(
  "index" = 1:10,
  "A" = sample(1:10),
  "B" = runif(10),
  "C" = LETTERS[1:10],
  stringsAsFactors = FALSE
)

# Center by the index (row numbers)
center_max(df)

# Center by each of the columns
center_max(df, col = "A")
center_max(df, col = "B")
center_max(df, col = "C")

# Randomize which elements are left and right of the center
center_max(df, col = "A", shuffle_members = TRUE)

# Plot the centered values
plot(x = 1:10, y = center_max(df, col = "B")$B)
plot(x = 1:10, y = center_max(df, col = "B", shuffle_members = TRUE)$B)
}
}
\seealso{
Other rearrange functions: 
\code{\link{center_min}()},
\code{\link{pair_extremes}()},
\code{\link{rearrange}()}
}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}
\concept{rearrange functions}
