% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearrange.R
\name{center_min}
\alias{center_min}
\title{Centers the lowest value with values increasing around it}
\usage{
center_min(data, col = NULL, shuffle_members = FALSE)
}
\arguments{
\item{data}{Data frame or vector. When a vector, it is converted to a data frame.}

\item{col}{Column to create sorting factor by. When \code{NULL} and \code{data} is a data frame,
it uses the row numbers.}

\item{shuffle_members}{Whether to shuffle the pair members. (Logical)

For the \code{"center_*"} methods, this randomizes which values are
to the right and left of the center.}
}
\value{
The sorted data frame.
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}

The lowest value is positioned in the middle with the other
values increasing around it.

Wrapper for \code{\link[groupdata2:rearrange]{rearrange()}} with \code{method = "center_min"}.

\strong{Example}:

The column values:

\code{c(}\strong{\code{1}}\code{, 2, 3, 4, 5)}

are \strong{ordered as}:

\code{c(5, 3,} \strong{\code{1}}\code{, 2, 4)}
}
\examples{
\donttest{
# Attach packages
library(groupdata2)
library(dplyr)

# Set seed
set.seed(1)

# Create a data frame
df <- data.frame(
  "index" = 1:10,
  "A" = sample(1:10),
  "B" = runif(10),
  "C" = LETTERS[1:10],
  "G" = c(1, 1, 1, 2, 2,
          2, 3, 3, 3, 3),
  stringsAsFactors = FALSE
)

# Center by the index (row numbers)
center_min(df)

# Center by each of the columns
center_min(df, col = "A")
center_min(df, col = "B")
center_min(df, col = "C")

# Randomize which elements are left and right of the center
center_min(df, col = "A", shuffle_members = TRUE)

# Grouped by G
df \%>\%
  dplyr::group_by(G) \%>\%
  center_min(col = "A")

# Plot the centered values
plot(x = 1:10, y = center_min(df, col = "B")$B)
plot(x = 1:10, y = center_min(df, col = "B", shuffle_members = TRUE)$B)
}
}
\seealso{
Other rearrange functions: 
\code{\link{center_max}()},
\code{\link{pair_extremes}()},
\code{\link{rearrange}()}
}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}
\concept{rearrange functions}
