% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{downsample}
\alias{downsample}
\title{Downsampling of rows in a dataframe.}
\usage{
downsample(data, cat_col)
}
\arguments{
\item{data}{Dataframe.}

\item{cat_col}{Name of categorical variable to balance by. (Character)}
}
\value{
Dataframe with some rows removed. Ordered by cat_col.
}
\description{
Uses random downsampling to fix the group sizes to the
 smallest group in the data frame.
}
\details{
Downsampling is done without replacement, meaning that rows are not duplicated but only removed.
}
\examples{
# Attach packages
library(groupdata2)

# Create dataframe
df <- data.frame(
 "participant" = factor(c(1, 1, 2, 3, 3, 3, 3)),
 "trial" = c(1,2,1,1,2,3,4),
 "score" = sample(c(1:100), 7))

# Using upsample() with number
downsample(df,"participant")

}
\seealso{
Other sampling functions: \code{\link{balance}},
  \code{\link{upsample}}
}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}
