% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primes_remainder.R
\name{\%primes\%}
\alias{\%primes\%}
\alias{primes}
\title{Find remainder from primes method.}
\usage{
size \%primes\% start_at
}
\arguments{
\item{size}{Size to group (Integer)}

\item{start_at}{Prime to start at (Integer)}
}
\value{
Remainder (Integer).
Returns \code{0} if the last group
has the size of the associated prime number.
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("stable")}

When using the primes method,
the last group might not have the size of the associated prime number
if there are not enough elements left. Use \code{\%primes\%} to find this remainder.
}
\examples{
# Attach packages
library(groupdata2)

100 \%primes\% 2
}
\seealso{
Other staircase tools: 
\code{\link{\%staircase\%}()},
\code{\link{group_factor}()},
\code{\link{group}()}

Other remainder tools: 
\code{\link{\%staircase\%}()}
}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}
\concept{remainder tools}
\concept{staircase tools}
