% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_info.R
\docType{package}
\name{groupdata2}
\alias{groupdata2}
\alias{groupdata2-package}
\title{groupdata2: A package for creating groups from data}
\description{
Subsetting Methods for Balanced Cross-Validation, Timeseries Windowing,
and General Grouping and Splitting of Data.
}
\details{
The groupdata2 package provides four main functions:
\code{group}, \code{group_factor}, \code{splt}, and \code{fold}
}
\section{group}{


Go to \code{\link{group}}
}

\section{group_factor}{


Go to \code{\link{group_factor}}
}

\section{splt}{


Go to \code{\link{splt}}
}

\section{fold}{


Go to \code{\link{fold}}
}
\author{
Ludvig Renbo Olsen, \email{mail@ludvigolsen.dk}
}

