% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearrange.R
\name{pair_extremes}
\alias{pair_extremes}
\title{Pair extreme values and sort by the pairs}
\usage{
pair_extremes(
  data,
  col = NULL,
  unequal_method = "middle",
  shuffle_members = FALSE,
  shuffle_pairs = FALSE,
  keep_factor = FALSE,
  factor_name = ".rearrange_factor"
)
}
\arguments{
\item{data}{Data frame or vector. When a vector, it is converted to a data frame.}

\item{col}{Column to create sorting factor by. When \code{NULL} and \code{data} is a data frame,
it uses the row numbers.}

\item{unequal_method}{Method for dealing with an unequal number of rows
in \code{data} when \code{method} is \code{"pair_extremes"}.

One of: \code{first}, \code{middle} or \code{last}

\subsection{first}{
The first group will have size \code{1}.

\strong{Example}:

The column values:

\code{c(1, 2, 3, 4, 5)}

Creates the \strong{sorting factor}:

\code{c(}\strong{\code{1}}\code{, 2, 3, 3, 2)}

And are \strong{ordered as}:

\code{c(}\strong{\code{1}}\code{, 2, 5, 3, 4)}

}

\subsection{middle}{
The middle group will have size \code{1}.

\strong{Example}:

The column values:

\code{c(1, 2, 3, 4, 5)}

Creates the \strong{sorting factor}:

\code{c(1, 3, }\strong{\code{2}}\code{, 3, 1)}

And are \strong{ordered as}:

\code{c(1, 5, } \strong{\code{3}}\code{, 2, 4)}

}
\subsection{last}{
The last group will have size \code{1}.

\strong{Example}:

The column values:

\code{c(1, 2, 3, 4, 5)}

Creates the \strong{sorting factor}:

\code{c(1, 2, 2, 1, }\strong{\code{3}}\code{)}

And are \strong{ordered as}:

\code{c(1, 4, 2, 3,} \strong{\code{5}}\code{)}

}}

\item{shuffle_members}{Whether to shuffle the pair members. (Logical)

For the \code{"center_*"} methods, this randomizes which values are
to the right and left of the center.}

\item{shuffle_pairs}{Whether to shuffle the pairs when
\code{method} is \code{"pair_extremes"}. (Logical)}

\item{keep_factor}{Whether to keep the sorting factor in the data frame. \code{Logical}.

This is mostly useful with the \code{"pair_extremes"} method.}

\item{factor_name}{Name of sorting factor.

N.B. Only used when \code{keep_factor} is \code{TRUE}.}
}
\value{
The sorted data frame. Optionally with the sorting factor added.
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}

The values are paired/grouped such that the highest and lowest values
form the first group, the second highest and the second lowest values
form the second group, and so on.
The values are then sorted by these groups/pairs.

When \code{data} has an uneven number of rows, the \code{unequal_method}
determines which group should have only \code{1} element.

Wrapper for \code{\link[groupdata2:rearrange]{rearrange()}} with \code{method = "pair_extremes"}.

\strong{Example}:

The column values:

\code{c(1, 2, 3, 4, 5, 6)}

Creates the \strong{sorting factor}:

\code{c(1, 2, 3, 3, 2, 1)}

And are \strong{ordered as}:

\code{c(1, 6, 2, 5, 3, 4)}
}
\examples{
\donttest{
# Attach packages
library(groupdata2)

# Set seed
set.seed(1)

# Create a data frame
df <- data.frame(
  "index" = 1:10,
  "A" = sample(1:10),
  "B" = runif(10),
  "C" = LETTERS[1:10],
  stringsAsFactors = FALSE
)

# Pair extreme indices (row numbers)
pair_extremes(df)

# Pair extremes in each of the columns
pair_extremes(df, col = "A")$A
pair_extremes(df, col = "B")$B
pair_extremes(df, col = "C")$C

# Shuffle the members pair-wise
pair_extremes(df, col = "A", shuffle_members = TRUE)

# Shuffle the order of the pairs
pair_extremes(df, col = "A", shuffle_pairs = TRUE)

# Plot the extreme pairs
plot(x = 1:10,
     y = pair_extremes(df, col = "B")$B)
# With shuffled pair members (run a few times)
plot(x = 1:10,
     y = pair_extremes(df, col = "B", shuffle_members = TRUE)$B)
# With shuffled pairs (run a few times)
plot(x = rep(1:5, each = 2),
     y = pair_extremes(df, col = "B", shuffle_pairs = TRUE)$B)
}
}
\seealso{
Other rearrange functions: 
\code{\link{center_max}()},
\code{\link{center_min}()},
\code{\link{rearrange}()}
}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}
\concept{rearrange functions}
