% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prime_numbers_description.R
\docType{data}
\name{prime_numbers}
\alias{prime_numbers}
\title{Prime numbers}
\format{A vector with 10,000,000 elements}
\usage{
prime_numbers
}
\description{
A vector with 10,000,000 prime numbers starting from 2.
}
\keyword{datasets}

