% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_info.R
\docType{package}
\name{splitters}
\alias{splitters}
\alias{splitters-package}
\title{splitters: A package for creating groups from data}
\description{
Create groups / windows from data with a range of methods.
Split data to a list or create balanced folds for cross-validation.
}
\details{
The splitters package provides four functions:
\code{group}, \code{group_factor}, \code{splt}, and \code{fold}
}
\section{group}{


Go to \code{\link{group}}
}

\section{group_factor}{


Go to \code{\link{group_factor}}
}

\section{splt}{


Go to \code{\link{splt}}
}

\section{fold}{


Go to \code{\link{fold}}
}

