% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{upsample}
\alias{upsample}
\title{Upsampling of rows in a dataframe.}
\usage{
upsample(data, cat_col)
}
\arguments{
\item{data}{Dataframe.}

\item{cat_col}{Name of categorical variable to balance by. (Character)}
}
\value{
Dataframe with added rows. Ordered by cat_col.
}
\description{
Uses random upsampling to fix the group sizes to the
 largest group in the data frame.
}
\details{
Upsampling is done with replacement for added rows, while the original data remains intact.
}
\examples{
# Attach packages
library(groupdata2)

# Create dataframe
df <- data.frame(
 "participant" = factor(c(1, 1, 2, 3, 3, 3, 3)),
 "trial" = c(1,2,1,1,2,3,4),
 "score" = sample(c(1:100), 7))

# Using upsample() with number
upsample(df,"participant")

}
\seealso{
Other sampling functions: \code{\link{balance}},
  \code{\link{downsample}}
}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}
