% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{balance}
\alias{balance}
\title{Balance groups by up- or downsampling.}
\usage{
balance(data, size, cat_col)
}
\arguments{
\item{data}{Dataframe.}

\item{size}{Size to fix group sizes to.
 Can be a specific number, given as a whole number, or one of the following strings:
 \code{"min"}, \code{"max"}, \code{"mean"}, \code{"median"}.

 \subsection{number}{
 Fix each group to have the size of the specified number of row.
 Uses downsampling for groups with too many rows and upsampling for groups with too few rows.
 }
 \subsection{min}{
 Fix each group to have the size of smallest group in the dataset.
 Uses downsampling on all groups that have too many rows.
 }
 \subsection{max}{
 Fix each group to have the size of largest group in the dataset.
 Uses upsampling on all groups that have too few rows.
 }
 \subsection{mean}{
 Fix each group to have the mean group size in the dataset. The mean is rounded.
 Uses downsampling for groups with too many rows and upsampling for groups with too few rows.
 }
 \subsection{median}{
 Fix each group to have the median group size in the dataset. The median is rounded.
 Uses downsampling for groups with too many rows and upsampling for groups with too few rows.
 }}

\item{cat_col}{Name of categorical variable to balance by. (Character)}
}
\value{
Dataframe with added and/or deleted rows. Ordered by cat_col.
}
\description{
Uses up- or downsampling to fix the group size to the
 min, max, mean, or median group size or to a specific number of rows.
}
\details{
Upsampling is done with replacement for added rows, while the original data remains intact.
Downsampling is done without replacement, meaning that rows are not duplicated but only removed.
}
\examples{
# Attach packages
library(groupdata2)

# Create dataframe
df <- data.frame(
 "participant" = factor(c(1, 1, 2, 3, 3, 3, 3)),
 "trial" = c(1,2,1,1,2,3,4),
 "score" = sample(c(1:100), 7))

# Using balance() with number
balance(df, 3, "participant")

# Using balance() with min
balance(df, "min", "participant")

# Using balance() with max
balance(df, "max", "participant")

}
\seealso{
Other sampling functions: \code{\link{downsample}},
  \code{\link{upsample}}
}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}
